+++ includes/admin/feedzy-rss-feeds-ui-lang.php
@@ -174,6 +174,12 @@ class Feedzy_Rss_Feeds_Ui_Lang {
 							),
 						),
 					),
+					'error_empty'       => array(
+						'label' => __( 'Message to show when feed is empty', 'feedzy-rss-feeds' ),
+						'placeholder' => __( 'Feed has no items.', 'feedzy-rss-feeds' ),
+						'type'        => 'text',
+						'value'       => '',
+					),
 				),
 			),
 			'section_item'  => array(
@@ -207,6 +213,21 @@ class Feedzy_Rss_Feeds_Ui_Lang {
 							'framename' => array(
 								'label' => __( 'framename', 'feedzy-rss-feeds' ),
 								'value' => 'framename',
+							),
+						),
+					),
+					'follow'         => array(
+						'label' => __( 'Make this link a "nofollow" link?', 'feedzy-rss-feeds' ),
+						'type'  => 'select',
+						'value' => '',
+						'opts'  => array(
+							'auto'      => array(
+								'label' => __( 'No', 'feedzy-rss-feeds' ),
+								'value' => '',
+							),
+							'_blank'    => array(
+								'label' => __( 'Yes', 'feedzy-rss-feeds' ),
+								'value' => 'no',
 							),
 						),
 					),
@@ -305,6 +326,25 @@ class Feedzy_Rss_Feeds_Ui_Lang {
 						'type'        => 'text',
 						'value'       => '',
 					),
+					'http'   => array(
+						'label' => __( 'How should we treat HTTP images?', 'feedzy-rss-feeds' ),
+						'type'  => 'select',
+						'value' => '',
+						'opts'  => array(
+							'auto' => array(
+								'label' => __( 'Show with HTTP link', 'feedzy-rss-feeds' ),
+								'value' => '',
+							),
+							'yes'  => array(
+								'label' => __( 'Force HTTPS (please verify that the images exist on HTTPS)', 'feedzy-rss-feeds' ),
+								'value' => 'force',
+							),
+							'no'   => array(
+								'label' => __( 'Ignore and show the default image instead', 'feedzy-rss-feeds' ),
+								'value' => 'default',
+							),
+						),
+					),
 				),
 			),
 			'section_pro'   => array(
@@ -346,6 +386,7 @@ class Feedzy_Rss_Feeds_Ui_Lang {
 						'type'     => 'text',
 						'disabled' => true,
 						'value'    => '',
+						'placeholder'    => '',
 					),
 					'template'     => array(
 						'label'    => __( 'Template to use when displaying the feed.', 'feedzy-rss-feeds' ),
@@ -394,5 +435,5 @@ class Feedzy_Rss_Feeds_Ui_Lang {
 
 }
 
-$feedzyLangClass = new Feedzy_Rss_Feeds_Ui_Lang();
-$strings         = $feedzyLangClass->feedzy_tinymce_translation();
+$feedzy_lang_class = new Feedzy_Rss_Feeds_Ui_Lang();
+$strings         = $feedzy_lang_class->feedzy_tinymce_translation();