+++ includes/layouts/feedzy-support.php
@@ -4,17 +4,18 @@
 
 	<?php
 	$active_tab  = isset( $_REQUEST['tab'] ) ? sanitize_text_field( $_REQUEST['tab'] ) : 'help';
+	$show_more = ! class_exists( 'Feedzy_Rss_Feeds_Pro' ) || false === apply_filters( 'feedzy_is_license_of_type', false, 'agency' );
 	?>
 
 	<h2 class="nav-tab-wrapper">
 		<a href="<?php echo esc_url( admin_url( 'admin.php?page=feedzy-support&tab=help' ) ); ?>"
-		   class="nav-tab <?php echo $active_tab == 'help' ? 'nav-tab-active' : ''; ?>"><?php _e( 'Support', 'feedzy-rss-feeds' ); ?></a>
+		   class="nav-tab <?php echo $active_tab === 'help' ? 'nav-tab-active' : ''; ?>"><?php _e( 'Support', 'feedzy-rss-feeds' ); ?></a>
 		<?php
-		if ( ! class_exists( 'Feedzy_Rss_Feeds_Pro' ) ) {
-		?>
+		if ( $show_more ) {
+			?>
 		<a href="<?php echo esc_url( admin_url( 'admin.php?page=feedzy-support&tab=more' ) ); ?>"
-	   class="nav-tab <?php echo $active_tab == 'more' ? 'nav-tab-active' : ''; ?>"><?php _e( 'More Features', 'feedzy-rss-feeds' ); ?></a>
-		<?php
+	   class="nav-tab <?php echo $active_tab === 'more' ? 'nav-tab-active' : ''; ?>"><?php _e( 'More Features', 'feedzy-rss-feeds' ); ?></a>
+			<?php
 		}
 		?>
 	</h2>
@@ -28,7 +29,7 @@
 							load_template( FEEDZY_ABSPATH . '/includes/layouts/feedzy-tutorial.php' );
 							break;
 						case 'more':
-							if ( ! class_exists( 'Feedzy_Rss_Feeds_Pro' ) ) {
+							if ( $show_more ) {
 								load_template( FEEDZY_ABSPATH . '/includes/layouts/feedzy-upsell.php' );
 							}
 							break;